Upgrade to ruby 2.4.4
Because Fixnum/Bignum were merged into an Integer 
type in 2.4 activesupport < 2.4.2 throw a stack 
level too deep error. Upgrading middleman brings
in a newer version of activesupport and fixes
this.

Upgrades to middleman also needed an upgrade to
middleman-livereload and middleman-compass.

I’m not exactly sure what the results of upgrading 
middleman-gh-pages is but I brought that up to
latest as well.

This still needs testing on windows as I have not
upgraded either of the gems for that platform. It
also needs testing with github deploys as I 
upgraded the gem but wasn’t able to test it.