Add back in sponsors content
When the original content was moved over here: 7b8051c47bd629cc7967af4c668469549400f18f it seems that the sponsors
content was dropped and pointed to program. Later it was pointed back
to sponsors but the partial was never added to file system.

Here we copied the content back as it was in that commit.